requirements.txt
Flask >= 1.0
gunicorn==20.1.0
google-cloud-storage==1.43.0
# panoptes-client>=1.5.0
git+https://github.com/zooniverse/panoptes-python-client.git
pg8000==1.23.0
SQLAlchemy==1.4.27
numpy>=1.19.5
psycopg2-binary
google-cloud-logging
# lsst-resources[s3]
# lsst-daf-butler
pytest
pytest-dotenv
pytest-cov